"素材库新建文件夹，移动，删除。消息队列，以及邮件"

+++ bkmanage/src/main/java/com/baojikouyu/teach/service/impl/FilesServiceImpl.java
@@ -6,12 +6,15 @@ import com.baojikouyu.teach.pojo.FilesDto;
 import com.baojikouyu.teach.service.FilesService;
 import com.baojikouyu.teach.util.AsyncTasks;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
+import com.baomidou.mybatisplus.core.conditions.update.LambdaUpdateWrapper;
 import com.baomidou.mybatisplus.core.toolkit.Wrappers;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import com.google.common.collect.Lists;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.annotation.CacheConfig;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
@@ -19,6 +22,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.interceptor.TransactionAspectSupport;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.Future;
@@ -30,6 +34,7 @@ import java.util.concurrent.Future;
  */
 @Service
 @Slf4j
+@CacheConfig(cacheNames = "files", keyGenerator = "keyGenerator")
 public class FilesServiceImpl extends ServiceImpl<FilesMapper, Files>
         implements FilesService {
 
@@ -40,22 +45,29 @@ public class FilesServiceImpl extends ServiceImpl<FilesMapper, Files>
     private AsyncTasks asyncTasks;
 
     @Override
-    @Cacheable(value = "getPageByType")
-    public Page<Files> getPageByType(Integer start, Integer size, Integer type) {
+//    @Cacheable(value = "getPageByType")
+    @Cacheable()
+    public Page<Files> getPageByType(Integer start, Integer size, Integer type, Integer grade, Integer parentFolderId) {
         LambdaQueryWrapper<Files> queryWrapper = Wrappers.lambdaQuery();
         Page<Files> page = new Page<>();
         page.setCurrent(start);
         page.setSize(size);
-        if (type != 0) {
+     /*   if (type != 0) {
             //判断是图片还是视频，还是文件
             queryWrapper.eq(Files::getType, type);
-        }
+        }*/
+        queryWrapper.eq(type != 0, Files::getType, type);
+        queryWrapper.eq(grade != null, Files::getFolderGrade, grade);
+        queryWrapper.eq(parentFolderId != null, Files::getParentFolderId, parentFolderId);
+        queryWrapper.orderByDesc(Files::getType, Files::getFolderGrade, Files::getUploadTime);
         return filesMapper.selectPage(page, queryWrapper);
     }
 
     @Override
     @Transactional
-    @CacheEvict(value = {"getPageByType", "getFileByNameAndSort"},
+//    @CacheEvict(value = {"getPageByType", "getFileByNameAndSort"},
+//            allEntries = true)
+    @CacheEvict(cacheNames = "files",
             allEntries = true)
     public Boolean saveFile(String filePath, String newFileName, MultipartFile file, Files files) {
         log.info("saveFile方法线程:{}", Thread.currentThread().getName());
@@ -98,9 +110,17 @@ public class FilesServiceImpl extends ServiceImpl<FilesMapper, Files>
     }
 
     @Override
+    @Cacheable
+    public Files getOneByParentFolderId(Integer parentFolderId) {
+        return filesMapper.selectOne(Wrappers.lambdaQuery(Files.class).eq(parentFolderId != null, Files::getId, parentFolderId));
+    }
+
+    @Override
     @Transactional
-    @CacheEvict(value = {"getPageByType", "getFileByNameAndSort"},
+    @CacheEvict(cacheNames = "files",
             allEntries = true)
+//    @CacheEvict(value = {"getPageByType", "getFileByNameAndSort"},
+//            allEntries = true)
     public Boolean deleteFilesByIds(Integer[] ids) {
         int i = 0;
         List<Files> files;
@@ -113,13 +133,16 @@ public class FilesServiceImpl extends ServiceImpl<FilesMapper, Files>
             return false;
         }
         files.forEach(files1 -> {
-            asyncTasks.deleteFile(files1.getFilePath() + files1.getUuidFileName());
+            if (!files1.getIsFolder()) {
+                asyncTasks.deleteFile(files1.getFilePath() + files1.getUuidFileName());
+            }
         });
         return i > 0;
     }
 
     @Override
-    @Cacheable(value = "getFileByNameAndSort")
+    @Cacheable
+//    @Cacheable(value = "getFileByNameAndSort")
     public Page<Files> getFileByNameAndSort(FilesDto fdto) {
         Page<Files> page = new Page<>(fdto.getStart() == 0 ? 0 : fdto.getStart(), fdto.getSize() == 0 ? 10 : fdto.getSize());
         LambdaQueryWrapper<Files> queryWrapper = Wrappers.lambdaQuery(Files.class);
@@ -133,6 +156,17 @@ public class FilesServiceImpl extends ServiceImpl<FilesMapper, Files>
             //0 仅自己创建的
             queryWrapper.eq(Files::getCreaterName, fdto.getCreater());
         }
+        if (fdto.getParentFolderId() != null) {
+            queryWrapper.eq(Files::getParentFolderId, fdto.getParentFolderId());
+        }
+        if (fdto.getFolderGrade() != null) {
+            queryWrapper.eq(Files::getFolderGrade, fdto.getFolderGrade());
+        }
+        // 排除文件夹
+        if (fdto.getExcludeFolder() != null && fdto.getExcludeFolder() == true) {
+            queryWrapper.eq(Files::getIsFolder, false);
+        }
+        queryWrapper.orderByDesc(Files::getType);
         if (fdto.getOrderValue() == 0) {
             //按照名称排序
             queryWrapper.orderByAsc(Files::getFileName);
@@ -140,7 +174,90 @@ public class FilesServiceImpl extends ServiceImpl<FilesMapper, Files>
             //按照时间排序
             queryWrapper.orderByAsc(Files::getUploadTime);
         }
+
         return filesMapper.selectPage(page, queryWrapper);
+    }
+
+    @Override
+    @CacheEvict(cacheNames = "files",
+            allEntries = true)
+    public Boolean saveFolder(Files files) {
+        int insert = filesMapper.insert(files);
+        if (insert > 0) {
+            return true;
+        } else return false;
+    }
+
+    @Override
+    @Cacheable
+    public List<Files> getAllFolder() {
+        LambdaQueryWrapper<Files> queryWrapper = Wrappers.lambdaQuery();
+        queryWrapper.eq(Files::getIsFolder, true);
+        return getTreeFolder(filesMapper.selectList(queryWrapper));
+    }
+
+    @Override
+    @Cacheable
+    public Boolean getFolderByParentFolderIdAndFileName(boolean b, Integer parentFolderId, String floderName) {
+        LambdaQueryWrapper<Files> queryWrapper = Wrappers.lambdaQuery();
+        queryWrapper.eq(Files::getIsFolder, true);
+        queryWrapper.eq(Files::getParentFolderId, parentFolderId);
+        queryWrapper.eq(Files::getFileName, floderName);
+
+        return filesMapper.exists(queryWrapper);
+
+    }
+
+    @Override
+    @CacheEvict(cacheNames = "files",
+            allEntries = true)
+    public Boolean updateByIds(Integer[] ids, Integer to, Integer grade) {
+        LambdaUpdateWrapper<Files> updateWrapper = Wrappers.lambdaUpdate(Files.class);
+        updateWrapper.in(ids.length > 0, Files::getId, ids);
+        updateWrapper.set(to != 0, Files::getParentFolderId, to);
+        updateWrapper.set(grade != null, Files::getFolderGrade, grade);
+        int update = filesMapper.update(null, updateWrapper);
+        if (update > 0) {
+            return true;
+        } else return false;
+    }
+
+    @Override
+    @Cacheable
+    public Files getOneById(Integer to) {
+        return filesMapper.selectOne(Wrappers.lambdaQuery(Files.class).eq(to != null, Files::getId, to));
+    }
+
+    @Override
+    @Cacheable
+    public List<Files> getBatchById(Integer[] ids) {
+        return filesMapper.selectBatchIds(Arrays.asList(ids));
+    }
+
+    @Override
+    @Cacheable
+    public List<Files> getChildrenByFolderParentId(Integer id) {
+        return filesMapper.selectList(Wrappers.lambdaQuery(Files.class).eq(id != null, Files::getParentFolderId, id));
+    }
+
+    public List<Files> getTreeFolder(List<Files> filesList) {
+        ArrayList<Files> finalFiles = Lists.newArrayList();
+        for (int i = 0; i < filesList.size(); i++) {
+
+            Files files = filesList.get(i);
+
+            for (int j = 0; j < filesList.size(); j++) {
+                Files tmp = filesList.get(j);
+                if (tmp.getParentFolderId() == files.getId()) {
+                    files.getChildrenFolder().add(tmp);
+                }
+            }
+            // 提取出父节点
+            if (files.getParentFolderId() == null || files.getParentFolderId() == 0) {
+                finalFiles.add(files);
+            }
+        }
+        return finalFiles;
     }
 }
